+++ Assets/Plugins/CountlySDK/Services/PushCountlyService.cs
@@ -117,12 +117,6 @@ namespace Plugins.CountlySDK.Services
         {
             EnableNotification();
         }
-
-        internal override void DeviceIdChanged(string deviceId, bool merged)
-        {
-
-        }
-
         internal override void ConsentChanged(List<Consents> updatedConsents, bool newConsentValue)
         {
             if (updatedConsents.Contains(Consents.Push) && newConsentValue && !_isDeviceRegistered) {