+++ Assets/Plugins/CountlySDK/Services/RemoteConfigCountlyService.cs
@@ -123,11 +123,6 @@ namespace Plugins.CountlySDK.Services
         }
 
         #region override Methods
-        internal override void DeviceIdChanged(string deviceId, bool merged)
-        {
-
-        }
-
         internal override void ConsentChanged(List<Consents> updatedConsents, bool newConsentValue)
         {
             if (updatedConsents.Contains(Consents.RemoteConfig) && !newConsentValue) {