+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * !! Major breaking change !! Changing device ID without merging will now clear the current consent. Consent has to be given again after performing this action.
 * 'UserCustomDetailAsync (CountlyUserDetailsModel userDetailsModel)' is deprecated.
 * Fixed issues with how parameter tampering protection was handled.
+* Device id can be changed without giving any consent.
 
 ## 20.11.5
 * Added new configuration fields to manipulate internal SDK value and key limits.